htop: use attrsOf type instead of attrs as the settings type
Read https://github.com/NixOS/nixpkgs/issues/85508 for why it's a bad idea to use the attrs type.
I noticed this problem while trying to use `mkDefault` for one of the options, which did not work and instead failed with the unhelpful error message `error: cannot coerce a set to a string`.
The new type is a little more strict than the old type. I could have probably used something like `attrsOf anything` but making the type a little more explicit seems like a good idea.